Move some stuff from floattypes.h into .c files
This makes it a bit easier to understand what is connected to what else.
