+++ src/floattypes.h
@@ -71,22 +71,6 @@ int InitMPFRLibrary (void);
  * mpfi
  ****************************************************************/
 #ifdef USE_MPFI
-#include <mpfi.h>
-
-/****************************************************************
- * mpfi's are stored as follows:
- * +-----------+-----------------------------------------+---------------------+
- * | TYPE_MPFI |             __mpfi_struct               |    __mp_limb_t[]    |
- * |           | __mpfr_struct left         right        | limbl ... limbr ... |
- * |           | prec exp sign mant   prec exp sign mant |                     |
- * +-----------+-----------------------------------------+---------------------+
- *                               \____________________\____^         ^
- *                                                     \____________/
- * it is assumed that the left and right mpfr's are allocated with the
- * same precision
- ****************************************************************/
-#define MPFI_OBJ(obj) ((mpfi_ptr) (ADDR_OBJ(obj)+1))
-
 int InitMPFIKernel (void);
 int InitMPFILibrary (void);
 #endif
@@ -95,8 +79,6 @@ int InitMPFILibrary (void);
  * mpc
  ****************************************************************/
 #ifdef USE_MPC
-#include <mpc.h>
-
 int InitMPCKernel (void);
 int InitMPCLibrary (void);
 #endif
@@ -121,52 +103,6 @@ int InitMPDLibrary (void);
  * cxsc
  ****************************************************************/
 #ifdef USE_CXSC
-#define ERROR_CXSC(gap_name,obj)				      \
-  ErrorQuit(#gap_name ": argument must be a CXSC float, not a %s",    \
-	    (Int)TNAM_OBJ(obj),0)
-
-#ifdef __cplusplus
-static inline bool HAS_FILTER(Obj obj, Obj filter)
-{
-  return DoFilter(filter,obj) == True;
-  return IS_DATOBJ(obj) && DoFilter(filter,obj) == True;
-}
-#endif
-#define IS_RP(obj) HAS_FILTER(obj,IS_CXSC_RP)
-#define TEST_IS_RP(gap_name,obj)				\
-  if (!IS_RP(obj))						\
-    ErrorQuit(#gap_name ": expected a real, not a %s",		\
-	       (Int)TNAM_OBJ(obj),0)
-
-#define IS_CP(obj) HAS_FILTER(obj,IS_CXSC_CP)
-#define TEST_IS_CP(gap_name,obj)				\
-  if (!IS_CP(obj))						\
-    ErrorQuit(#gap_name ": expected a complex, not a %s",	\
-	       (Int)TNAM_OBJ(obj),0)
-
-#define IS_RI(obj) HAS_FILTER(obj,IS_CXSC_RI)
-#define TEST_IS_RI(gap_name,obj)			       	\
-  if (!IS_RI(obj))						\
-    ErrorQuit(#gap_name ": expected an interval, not a %s",	\
-	       (Int)TNAM_OBJ(obj),0)
-
-#define IS_CI(obj) HAS_FILTER(obj,IS_CXSC_CI)
-#define TEST_IS_CI(gap_name,obj)			       	\
-  if (!IS_CI(obj))					       	\
-    ErrorQuit(#gap_name ": expected a complex interval, not a %s",\
-	       (Int)TNAM_OBJ(obj),0)
-
-/****************************************************************
- * cxsc data are stored as follows:
- * +--------------------+----------+
- * | TYPE_CXSC_RI       | interval  |
- * +--------------------+----------+
- ****************************************************************/
-#define RP_OBJ(obj) (*(cxsc::real *) (ADDR_OBJ(obj)+1))
-#define RI_OBJ(obj) (*(cxsc::interval *) (ADDR_OBJ(obj)+1))
-#define CP_OBJ(obj) (*(cxsc::complex *) (ADDR_OBJ(obj)+1))
-#define CI_OBJ(obj) (*(cxsc::cinterval *) (ADDR_OBJ(obj)+1))
-
 int InitCXSCKernel (void);
 int InitCXSCLibrary (void);
 #endif
