+++ src/mpc.c
@@ -21,6 +21,8 @@
 
 #include "floattypes.h"
 
+#include <mpc.h>
+
 /****************************************************************
  * mpc's are stored as follows:
  * +----------+-----------------------------------------+---------------------+
